Move container crate to projectcontainers app
